Refactor component folder structure and update imports
