+++ client/src/pages/dashboard/components/DocumentSearchBar.tsx
