+++ client/src/pages/login/index.tsx
@@ -1,13 +1,13 @@
-import useWindowSize from '../../hooks/useWindowSize';
-import TextField from '../../components/TextField';
+import useWindowSize from '@src/hooks/useWindowSize';
+import TextField from '@src/components/TextField';
 import { type KeyboardEvent, useContext, useState } from 'react';
-import { ToastContext } from '../../context/ToastContext';
-import Logo from '../../components/Logo';
+import { ToastContext } from '@src/context/ToastContext';
+import Logo from '@src/components/Logo';
 import validator from 'validator';
-import Spinner from '../../components/Spinner';
+import Spinner from '@src/components/Spinner';
 import { Link, useNavigate } from 'react-router-dom';
-import useAuth from '../../hooks/useAuth';
-import AuthService from '../../services/auth-service';
+import useAuth from '@src/hooks/useAuth';
+import AuthService from '@src/services/auth-service';
 import axios, { type AxiosError } from 'axios';
 
 const LoginPage = (): JSX.Element => {
